+++ packages/bluesky/lib/src/services/codegen/app/bsky/unspecced/getTrends/output.g.dart
@@ -21,6 +21,7 @@ _UnspeccedGetTrendsOutput _$UnspeccedGetTrendsOutputFromJson(Map json) =>
               )
               .toList(),
         ),
+        recIdStr: $checkedConvert('recIdStr', (v) => v as String?),
         $unknown: $checkedConvert(
           r'$unknown',
           (v) => (v as Map?)?.map((k, e) => MapEntry(k as String, e)),
@@ -33,5 +34,6 @@ Map<String, dynamic> _$UnspeccedGetTrendsOutputToJson(
   _UnspeccedGetTrendsOutput instance,
 ) => <String, dynamic>{
   'trends': instance.trends.map(const TrendViewConverter().toJson).toList(),
+  'recIdStr': ?instance.recIdStr,
   r'$unknown': ?instance.$unknown,
 };