+++ packages/bluesky/lib/src/services/codegen/app/bsky/unspecced/getTrendsSkeleton/output.g.dart
@@ -24,6 +24,7 @@ _UnspeccedGetTrendsSkeletonOutput _$UnspeccedGetTrendsSkeletonOutputFromJson(
           )
           .toList(),
     ),
+    recIdStr: $checkedConvert('recIdStr', (v) => v as String?),
     $unknown: $checkedConvert(
       r'$unknown',
       (v) => (v as Map?)?.map((k, e) => MapEntry(k as String, e)),
@@ -36,5 +37,6 @@ Map<String, dynamic> _$UnspeccedGetTrendsSkeletonOutputToJson(
   _UnspeccedGetTrendsSkeletonOutput instance,
 ) => <String, dynamic>{
   'trends': instance.trends.map(const SkeletonTrendConverter().toJson).toList(),
+  'recIdStr': ?instance.recIdStr,
   r'$unknown': ?instance.$unknown,
 };