+++ packages/bluesky_cli/lib/src/commands/codegen/app/bsky/graph.dart
@@ -36,6 +36,7 @@ import 'graph/mute_actor.dart';
 import 'graph/mute_actor_list.dart';
 import 'graph/mute_thread.dart';
 import 'graph/search_starter_packs.dart';
+import 'graph/search_starter_packs_v2.dart';
 import 'graph/starterpack.dart';
 import 'graph/unmute_actor.dart';
 import 'graph/unmute_actor_list.dart';
@@ -73,6 +74,7 @@ final class AppBskyGraphCommand extends Command<void> {
     addSubcommand(MuteActorListCommand());
     addSubcommand(MuteThreadCommand());
     addSubcommand(SearchStarterPacksCommand());
+    addSubcommand(SearchStarterPacksV2Command());
     addSubcommand(StarterpackCommand());
     addSubcommand(UnmuteActorCommand());
     addSubcommand(UnmuteActorListCommand());