+++ packages/bluesky_cli/lib/src/commands/codegen/app/bsky/graph/search_starter_packs_v2.dart
@@ -1,0 +1,52 @@
+// Copyright (c) 2023-2026, Shinya Kato.
+// All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+// Project imports:
+import '../../../../query_command.dart';
+
+// **************************************************************************
+// LexGenerator
+// **************************************************************************
+
+final class SearchStarterPacksV2Command extends QueryCommand {
+  SearchStarterPacksV2Command() {
+    argParser
+      ..addOption(
+        "q",
+        help:
+            r"Search query string. Syntax, phrase, boolean, and faceting is unspecified, but Lucene query syntax is recommended.",
+        mandatory: true,
+      )
+      ..addOption("limit", defaultsTo: "25")
+      ..addOption("cursor");
+  }
+
+  @override
+  final String name = "search-starter-packs-v2";
+
+  @override
+  final String description =
+      "Find starter packs matching search criteria. Does not require auth.";
+
+  @override
+  final String invocation =
+      "bsky app-bsky-graph search-starter-packs-v2 --q=<value> [--limit=<value>] [--cursor=<value>]";
+
+  @override
+  String get methodId => "app.bsky.graph.searchStarterPacksV2";
+
+  @override
+  Map<String, dynamic>? get parameters => {
+    "q": argResults!["q"],
+    "limit":
+        int.tryParse(argResults!["limit"]) ??
+        usageException('Invalid integer value for option "limit".'),
+    if (argResults!.wasParsed("cursor")) "cursor": argResults!["cursor"],
+  };
+}